+++ NEWS.md
@@ -1,3 +1,7 @@
+# dgrng (unreleased)
+
+* new method `dqrrademacher` for drawing Rademacher weights (Kyle Butts in [#50](https://github.com/daqana/dqrng/pull/50 fixing [#49](https://github.com/daqana/dqrng/pull/49))
+
 # dqrng 0.3.0
 
 ## Breaking changes