+++ NEWS.md
@@ -1,4 +1,4 @@
-# dgrng 0.3.1
+# dqrng 0.3.1
 
 * new method `dqrrademacher` for drawing Rademacher weights (Kyle Butts in [#50](https://github.com/daqana/dqrng/pull/50) fixing [#49](https://github.com/daqana/dqrng/pull/49))
 * Move sampling methods to separate header file, allowing for parallel usage.